Enable SSL/TLS via properties secure and securePort
Add two new properties, "secure" and "securePort" to control whether the
connection is made over SSL/TLS. This is an alternative to the
previously proposed implementation where scheme and port were embedded
in the URL and avoids breaking compatibility with existing connectors.
GeoEvent may need to be restarted when deploying this change, as content
of properties files may be cached. (Although not directly related, the
default values for host and topic are also changed in the XML transport
definitions to match the Java source code. This does not affect existing
connectors.)